# [Core][Model] Use torch.compile to accelerate layernorm in commandr